feat(ios): extract message bubble into reusable component
